backend/requirements.txt
Flask
python-dotenv
pandas
scikit-learn
matplotlib
seaborn
ipywidgets
flask-cors
gunicorn
scikit-surprise
python-dotenv
pyarrow
fastparquet
redis